move toxicity classifier creation to model_load
